fix bug 1470895: move siglists/ under socorro/signature/
The siglists are used by the signature generator and only the signature
generator. To make the signature generator a self-contained module, we
need to move the siglists folder under it.
