+++ socorro/signature/README.rst
@@ -57,7 +57,6 @@ This code can sort of be used as a library. It's been decoupled from many of
 Socorro's bits, but still has some requirements. Roughtly, it requires:
 
 * requests
-* socorro.siglists
 * ujson
