+++ socorro/signature/siglists/irrelevant_signature_re.txt
